refactor: remove phone, gender, birth
