--- core/auth/src/main/kotlin/com/pida/authentication/domain/token/PhoneToken.kt
@@ -1,5 +1,0 @@
-package com.pida.authentication.domain.token
-
-data class PhoneToken(
-    val token: String,
-)
