+++ core/core-api/src/main/kotlin/com/pida/presentation/v1/auth/request/SignUpSocialRequest.kt
@@ -4,17 +4,12 @@ import com.pida.authentication.domain.auth.AuthorityType
 import com.pida.authentication.domain.auth.GrantedAuthority
 import com.pida.authentication.domain.auth.NewAuthenticationSocial
 import com.pida.authentication.domain.auth.SocialType
-import com.pida.user.Gender
 import com.pida.user.NewUser
-import java.time.LocalDate
 
 data class SignUpSocialRequest(
     val email: String,
     val name: String,
     val nickName: String,
-    val phone: String,
-    val gender: Gender,
-    val birth: LocalDate,
     val socialId: String,
     val socialToken: String,
     val socialType: SocialType,
@@ -24,9 +19,6 @@ data class SignUpSocialRequest(
             email = email,
             name = name,
             nickname = nickName,
-            phone = phone,
-            gender = gender,
-            birth = birth,
         )
 
     fun toNewAuthenticationSocial(): NewAuthenticationSocial =
