+++ core/core-domain/src/main/kotlin/com/pida/user/AppleUserInfo.kt
@@ -1,0 +1,7 @@
+package com.pida.user
+
+data class AppleUserInfo(
+    val id: String,
+    val email: String,
+    val name: String,
+)
