+++ core/core-domain/src/main/kotlin/com/pida/user/UserProfile.kt
@@ -1,6 +1,5 @@
 package com.pida.user
 
-import java.time.LocalDate
 import java.time.LocalDateTime
 
 data class UserProfile(
@@ -9,8 +8,5 @@ data class UserProfile(
     val email: String,
     val name: String,
     val nickname: String,
-    val phone: String,
-    val gender: Gender,
-    val birth: LocalDate,
     val createdAt: LocalDateTime,
 )
