+++ core/core-domain/src/main/kotlin/com/pida/user/UserService.kt
@@ -11,7 +11,7 @@ class UserService(
     private val userValidator: UserValidator,
 ) {
     suspend fun create(newUser: NewUser): User {
-        userValidator.verify(newUser)
+        userValidator.verifyEmail(newUser.email)
         return userAppender.create(newUser)
     }
 
@@ -19,18 +19,26 @@ class UserService(
 
     fun getUser(userId: Long): User = userReader.readUser(userId)
 
+    fun signInKakao(kakaoUserInfo: KakaoUserInfo): User {
+        // 있으면 바로 return
+        val user = userReader.readUserByEmail(kakaoUserInfo.email)
+        return if (user != null) {
+            user
+        } else {
+            val newUser =
+                NewUser(
+                    name = kakaoUserInfo.name,
+                    email = kakaoUserInfo.email,
+                    nickname = kakaoUserInfo.nickname,
+                )
+            userAppender.create(newUser)
+        }
+    }
+
     suspend fun updateNickname(
         userKey: String,
         updateNickname: UpdateNickname,
     ): UserProfile = userUpdater.updateNickname(userKey, updateNickname)
-
-    suspend fun updatePhone(
-        userKey: String,
-        phone: String,
-    ): UserProfile {
-        userValidator.verifyPhone(phone)
-        return userUpdater.updatePhone(userKey, phone)
-    }
 
     suspend fun updateEmail(
         userKey: String,
@@ -41,26 +49,9 @@ class UserService(
         userDeleter.deleteUser(userKey)
     }
 
-    suspend fun getUser(
-        name: String,
-        phone: String,
-    ): UserProfile = userReader.readUser(name, phone)
-
-    suspend fun getUserProfile(
-        name: String,
-        phone: String,
-    ): UserProfile {
-        val userProfile = userReader.readUserProfile(name, phone)
-        return userProfile
-    }
-
     suspend fun getAllUserProfile(userIds: List<Long>): List<UserProfile> = userReader.readAllByUserIds(userIds)
 
     suspend fun checkEmail(email: String) {
         userValidator.verifyEmail(email)
-    }
-
-    suspend fun verifyUser(validateNewUser: ValidateNewUser) {
-        userValidator.verifyPhone(validateNewUser.phone)
     }
 }
