+++ core/core-domain/src/main/kotlin/com/pida/user/UserValidator.kt
@@ -6,21 +6,8 @@ import org.springframework.stereotype.Component
 
 @Component
 class UserValidator(
-    private val userReader: UserReader,
     private val userRepository: UserRepository,
 ) {
-    suspend fun verify(newUser: NewUser) {
-        if (userReader.existsByEmailOrPhone(newUser.email, newUser.phone)) {
-            throw ErrorException(ErrorType.DUPLICATED_USER)
-        }
-    }
-
-    suspend fun verifyPhone(phone: String) {
-        if (userRepository.existsByPhone(phone)) {
-            throw ErrorException(ErrorType.DUPLICATED_PHONE)
-        }
-    }
-
     suspend fun verifyEmail(email: String) {
         if (userRepository.existsByEmail(email)) {
             throw ErrorException(ErrorType.DUPLICATED_EMAIL)
