+++ core/core-domain/src/main/kotlin/com/pida/user/ValidateNewUser.kt
@@ -1,10 +1,5 @@
 package com.pida.user
 
-import java.time.LocalDate
-
 data class ValidateNewUser(
     val name: String,
-    val birth: LocalDate,
-    val gender: Gender,
-    val phone: String,
 )
