+++ storage/db-core/src/main/kotlin/com/pida/storage/db/core/user/UserEntity.kt
@@ -1,17 +1,13 @@
 package com.pida.storage.db.core.user
 
 import com.pida.storage.db.core.support.BaseEntity
-import com.pida.user.Gender
 import com.pida.user.NewUser
 import com.pida.user.NewUserKey
 import com.pida.user.User
 import com.pida.user.UserProfile
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
-import jakarta.persistence.EnumType
-import jakarta.persistence.Enumerated
 import jakarta.persistence.Table
-import java.time.LocalDate
 
 @Entity
 @Table(name = "t_users")
@@ -21,11 +17,6 @@ class UserEntity(
     var name: String,
     var nickname: String,
     var email: String,
-    var phone: String,
-    @Enumerated(value = EnumType.STRING)
-    @Column(columnDefinition = "varchar(50)")
-    val gender: Gender,
-    var birth: LocalDate,
 ) : BaseEntity() {
     constructor(
         newUser: NewUser,
@@ -35,9 +26,6 @@ class UserEntity(
         email = newUser.email,
         name = newUser.name,
         nickname = newUser.nickname,
-        phone = newUser.phone,
-        gender = newUser.gender,
-        birth = newUser.birth,
     )
 
     fun toUser(): User =
@@ -53,18 +41,11 @@ class UserEntity(
             email = email,
             name = name,
             nickname = nickname,
-            phone = phone,
-            gender = gender,
-            birth = birth,
             createdAt = createdAt,
         )
 
     fun updateNickname(nickname: String) {
         this.nickname = nickname
-    }
-
-    fun updatePhone(phone: String) {
-        this.phone = phone
     }
 
     fun updateEmail(email: String) {
