+++ storage/db-core/src/main/kotlin/com/pida/storage/db/core/user/UserJpaRepository.kt
@@ -8,25 +8,13 @@ interface UserJpaRepository :
     KotlinJdslJpqlExecutor {
     fun findByUserKey(userKey: String): UserEntity?
 
-    fun findByNameAndPhoneAndDeletedAtIsNull(
-        name: String,
-        phone: String,
-    ): UserEntity?
-
-    fun findByPhone(phone: String): UserEntity?
-
     fun findAllByIdIn(ids: List<Long>): List<UserEntity>
 
     fun findByIdAndDeletedAtIsNull(id: Long): UserEntity?
 
-    fun existsByEmailAndPhoneAndDeletedAtIsNull(
-        email: String,
-        phone: String,
-    ): Boolean
-
     fun existsByEmailAndDeletedAtIsNull(email: String): Boolean
 
-    fun existsByPhoneAndDeletedAtIsNull(phone: String): Boolean
-
     fun deleteByUserKey(userKey: String)
+
+    fun findByEmail(email: String): UserEntity?
 }
